+++ aragog_cli/src/config.rs
@@ -1,5 +1,4 @@
-use clap::ArgMatches;
-
+use crate::app::AragogCliApp;
 use crate::error::AragogCliError;
 use crate::log_level::LogLevel;
 use std::fmt::{self, Display, Formatter};
@@ -28,43 +27,43 @@ pub fn log<T: Display>(text: T, level: LogLevel) {
 }
 
 impl Config {
-    pub fn new(matches: &ArgMatches) -> Result<Self, AragogCliError> {
+    pub fn new(opts: &AragogCliApp) -> Result<Self, AragogCliError> {
         unsafe {
-            LOG_LEVEL = LogLevel::from(matches.occurrences_of("verbose"));
+            LOG_LEVEL = LogLevel::from(opts.verbose);
             log(format!("Log level: {:?}", LOG_LEVEL), LogLevel::Verbose);
         }
         let res = Self {
-            schema_collection_name: matches
-                .value_of("aragog_collection")
-                .unwrap_or(ARAGOG_DEFAULT_COLLECTION)
-                .to_string(),
+            schema_collection_name: opts
+                .schema_collection_name
+                .clone()
+                .unwrap_or_else(|| ARAGOG_DEFAULT_COLLECTION.to_string()),
             schema_path: {
-                match Self::load_str(matches, "schema_path", "SCHEMA_PATH", "path") {
+                match Self::load_str(&opts.schema_path, "schema_path", "SCHEMA_PATH", "path") {
                     Ok(val) => val,
                     Err(_err) => String::from(aragog::schema::SCHEMA_DEFAULT_PATH),
                 }
             },
-            db_host: Self::load_str(matches, "db_host", "DB_HOST", "db-host")?,
-            db_name: Self::load_str(matches, "db_name", "DB_NAME", "db-name")?,
-            db_user: Self::load_str(matches, "db_user", "DB_USER", "db-user")?,
-            db_pwd: Self::load_str(matches, "db_password", "DB_PASSWORD", "db-password")?,
+            db_host: Self::load_str(&opts.db_host, "db_host", "DB_HOST", "db-host")?,
+            db_name: Self::load_str(&opts.db_name, "db_name", "DB_NAME", "db-name")?,
+            db_user: Self::load_str(&opts.db_user, "db_user", "DB_USER", "db-user")?,
+            db_pwd: Self::load_str(&opts.db_pwd, "db_password", "DB_PASSWORD", "db-password")?,
         };
         log(&res, LogLevel::Verbose);
         Ok(res)
     }
 
     pub fn load_str(
-        matches: &ArgMatches,
-        value: &str,
+        value: &Option<String>,
+        value_name: &str,
         env_default: &str,
         option: &str,
     ) -> Result<String, AragogCliError> {
-        match matches.value_of(value) {
+        match value {
             Some(value) => Ok(value.to_string()),
             None => match std::env::var(env_default) {
                 Ok(value) => Ok(value),
                 Err(_error) => Err(AragogCliError::InitError {
-                    item: value.to_string(),
+                    item: value_name.to_string(),
                     message: format!(
                         "{} is not specified, please set the env var or use the --{} option",
                         env_default, option
@@ -99,23 +98,33 @@ impl Display for Config {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::app::Command;
 
     #[test]
     fn default_values_work() {
-        let matches = ArgMatches::default();
+        let opts = AragogCliApp {
+            schema_collection_name: None,
+            schema_path: None,
+            db_host: None,
+            db_name: None,
+            db_user: None,
+            db_pwd: None,
+            verbose: 0,
+            command: Command::Migrate,
+        };
         std::env::set_var("DB_HOST", "test_DB_HOST");
         std::env::set_var("DB_NAME", "test_DB_NAME");
         std::env::set_var("DB_USER", "test_DB_USER");
         std::env::set_var("DB_PASSWORD", "test_DB_PASSWORD");
         std::env::set_var("SCHEMA_PATH", "test_path");
-        let config = Config::new(&matches).unwrap();
+        let config = Config::new(&opts).unwrap();
         assert_eq!(config.db_host, "test_DB_HOST".to_string());
         assert_eq!(config.db_name, "test_DB_NAME".to_string());
         assert_eq!(config.db_user, "test_DB_USER".to_string());
         assert_eq!(config.db_pwd, "test_DB_PASSWORD".to_string());
         assert_eq!(config.schema_path, "test_path".to_string());
         std::env::remove_var("SCHEMA_PATH");
-        let config = Config::new(&matches).unwrap();
+        let config = Config::new(&opts).unwrap();
         assert_eq!(
             config.schema_path,
             aragog::schema::SCHEMA_DEFAULT_PATH.to_string()